+++ src/app.module.ts
@@ -26,7 +26,7 @@ import { AdminCommunicationModule } from '@platform/admin/communication/admin.co
 import { AppController } from '@src/app.controller';
 import { WinstonConfigService } from '@src/config/winston.config';
 import { MetadataModule } from '@src/platform/metadata/metadata.module';
-import { SearchModule } from '@services/api/search/search.module';
+import { SearchModule } from '@services/api/search/v1/search.module';
 import { KonfigModule } from '@src/platform/configuration/config/config.module';
 import { print } from 'graphql/language/printer';
 import { WinstonModule } from 'nest-winston';
