+++ src/platform/admin/search/admin.search.ingest.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AdminSearchIngestResult } from './admin.search.ingest.result';
-import { SearchIngestModule } from '@services/api/search2/search.ingest/search.ingest.module';
+import { SearchIngestModule } from '@services/api/search/v2/ingest';
 import { AdminSearchIngestResolverMutations } from '@platform/admin/search/admin.search.ingest.resolver.mutations';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { PlatformAuthorizationPolicyModule } from '@platform/authorization/platform.authorization.policy.module';
