+++ src/services/api/search/dto/index.ts
@@ -1,0 +1,12 @@
+export * from './search.dto.input';
+export * from './search.result.dto';
+export * from './search.result.dto.entry.post';
+export * from './search.result.dto.entry.user';
+export * from './search.result.dto.entry.space';
+export * from './search.result.entry.interface';
+export * from './search.result.dto.entry.callout';
+export * from './search.result.dto.entry.challenge';
+export * from './search.result.dto.entry.user.group';
+export * from './search.result.dto.entry.opportunity';
+export * from './search.result.dto.entry.organization';
+export * from './search.result.dto.entry.base.interface';
