--- src/services/api/search/index.ts
@@ -1,2 +1,0 @@
-export * from './dto/search.dto.input';
-export * from './search.service';
