+++ src/services/api/search/search.entity.types.ts
@@ -1,5 +1,6 @@
 export enum SearchEntityTypes {
   USER = 'user',
+  GROUP = 'group', // may be removed
   ORGANIZATION = 'organization',
   SPACE = 'space',
   SUBSPACE = 'subspace',
