+++ src/services/api/search/v1/index.ts
@@ -1,0 +1,2 @@
+export * from '@services/api/search/dto/search.dto.input';
+export * from './search.service';
