+++ src/services/api/search/v1/search.resolver.queries.ts
@@ -1,12 +1,12 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Resolver, Query } from '@nestjs/graphql';
 import { CurrentUser, Profiling } from '@src/common/decorators';
-import { SearchInput } from './dto/search.dto.input';
+import { SearchInput } from '@services/api/search/dto/search.dto.input';
 import { GraphqlGuard } from '@core/authorization';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
-import { ISearchResults } from './dto/search.result.dto';
+import { ISearchResults } from '@services/api/search/dto/search.result.dto';
 import { SearchService } from './search.service';
-import { Search2Service } from '../search2/search2.service';
+import { Search2Service } from '@services/api/search/v2/search2.service';
 import { ConfigService } from '@nestjs/config';
 import { ConfigurationTypes } from '@common/enums';
 @Resolver()
