+++ src/services/api/search/v1/search.result.builder.interface.ts
@@ -1,13 +1,13 @@
-import { ISearchResult } from './dto/search.result.entry.interface';
+import { ISearchResult } from '@services/api/search/dto/search.result.entry.interface';
 import { SearchResultType } from '@common/enums/search.result.type';
-import { ISearchResultSpace } from './dto/search.result.dto.entry.space';
-import { ISearchResultChallenge } from './dto/search.result.dto.entry.challenge';
-import { ISearchResultOpportunity } from './dto/search.result.dto.entry.opportunity';
-import { ISearchResultUser } from './dto/search.result.dto.entry.user';
-import { ISearchResultOrganization } from './dto/search.result.dto.entry.organization';
-import { ISearchResultUserGroup } from './dto/search.result.dto.entry.user.group';
-import { ISearchResultPost } from './dto/search.result.dto.entry.post';
-import { ISearchResultCallout } from './dto/search.result.dto.entry.callout';
+import { ISearchResultSpace } from '@services/api/search/dto/search.result.dto.entry.space';
+import { ISearchResultChallenge } from '@services/api/search/dto/search.result.dto.entry.challenge';
+import { ISearchResultOpportunity } from '@services/api/search/dto/search.result.dto.entry.opportunity';
+import { ISearchResultUser } from '@services/api/search/dto/search.result.dto.entry.user';
+import { ISearchResultOrganization } from '@services/api/search/dto/search.result.dto.entry.organization';
+import { ISearchResultUserGroup } from '@services/api/search/dto/search.result.dto.entry.user.group';
+import { ISearchResultPost } from '@services/api/search/dto/search.result.dto.entry.post';
+import { ISearchResultCallout } from '@services/api/search/dto/search.result.dto.entry.callout';
 
 interface SearchResultBuilderFunction<TypedSearchResult> {
   (rawSearchResult: ISearchResult): Promise<TypedSearchResult>;
