+++ src/services/api/search/v2/extract/function.score.functions.ts
