+++ src/services/api/search/v2/ingest/get.index.pattern.ts
