+++ src/services/api/search/v2/ingest/index.ts
@@ -1,0 +1,1 @@
+export * from './search.ingest.module';
