+++ test/mocks/search/search.extract.service.mock.ts
@@ -1,6 +1,6 @@
 import { ValueProvider } from '@nestjs/common';
 import { PublicPart } from '@test/utils';
-import { SearchExtractService } from '@services/api/search2/search.extract.service';
+import { SearchExtractService } from '@services/api/v2/search.extract.service';
 
 export const MockSearchExtractService: ValueProvider<
   PublicPart<SearchExtractService>
